requirements.txt
# Maritime Scraper Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
